pyproject.toml
[project]
name = "overleaf-cli"
version = "0.1.0"
description = "CLI tool for bidirectional sync with Overleaf"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "beautifulsoup4>=4.11",
    "browser-cookie3>=0.19",
    "websocket-client>=1.6",
]

[project.scripts]
overleaf = "overleaf_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/overleaf_cli"]
